+++ ThirdPartyAdapters/chartboost/chartboost/src/main/java/com/google/ads/mediation/chartboost/ChartboostMediationAdapter.java
@@ -4,6 +4,8 @@ import android.content.Context;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
+import androidx.annotation.IntDef;
+import androidx.annotation.NonNull;
 import com.chartboost.sdk.Chartboost;
 import com.chartboost.sdk.Model.CBError;
 import com.google.android.gms.ads.mediation.Adapter;
@@ -14,12 +16,27 @@ import com.google.android.gms.ads.mediation.MediationRewardedAd;
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback;
 import com.google.android.gms.ads.mediation.MediationRewardedAdConfiguration;
 import com.google.android.gms.ads.mediation.VersionInfo;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 import java.util.HashMap;
 import java.util.List;
 
 public class ChartboostMediationAdapter extends Adapter implements MediationRewardedAd {
 
   static final String TAG = ChartboostMediationAdapter.class.getSimpleName();
+
+  // region Error codes
+  /** Chartboost adapter errors. */
+  @Retention(RetentionPolicy.SOURCE)
+  @IntDef(value = {ERROR_BANNER_SIZE_MISMATCH, ERROR_AD_ALREADY_LOADED})
+  public @interface AdapterError {}
+
+  /** The requested ad size does not match a Chartboost supported banner size. */
+  static final int ERROR_BANNER_SIZE_MISMATCH = 100;
+
+  /** Chartboost can only load 1 ad per location at a time. */
+  static final int ERROR_AD_ALREADY_LOADED = 101;
+  // endregion
 
   /** A Chartboost extras object used to store optional information used when loading ads. */
   private ChartboostParams mChartboostParams = new ChartboostParams();
@@ -48,7 +65,7 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
 
     String logMessage =
         String.format(
-            "Unexpected adapter version format: %s." + "Returning 0.0.0 for adapter version.",
+            "Unexpected adapter version format: %s. Returning 0.0.0 for adapter version.",
             versionString);
     Log.w(TAG, logMessage);
     return new VersionInfo(0, 0, 0);
@@ -68,8 +85,7 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
 
     String logMessage =
         String.format(
-            "Unexpected SDK version format: %s." + "Returning 0.0.0 for SDK version.",
-            versionString);
+            "Unexpected SDK version format: %s. Returning 0.0.0 for SDK version.", versionString);
     Log.w(TAG, logMessage);
     return new VersionInfo(0, 0, 0);
   }
@@ -129,8 +145,6 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
     final Bundle serverParameters = mediationRewardedAdConfiguration.getServerParameters();
     final Bundle extras = mediationRewardedAdConfiguration.getMediationExtras();
 
-    Context context = mediationRewardedAdConfiguration.getContext();
-
     mChartboostParams = ChartboostAdapterUtils.createChartboostParams(serverParameters, extras);
     if (!ChartboostAdapterUtils.isValidChartboostParams(mChartboostParams)) {
       // Invalid server parameters, send initialization failed event.
@@ -140,14 +154,7 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
       return;
     }
 
-    if (!ChartboostAdapterUtils.isValidContext(context)) {
-      // Chartboost initialization failed, send initialization failed event.
-      String logMessage = "Failed to request ad from Chartboost: Internal Error.";
-      Log.w(TAG, logMessage);
-      mediationAdLoadCallback.onFailure(logMessage);
-      return;
-    }
-
+    Context context = mediationRewardedAdConfiguration.getContext();
     ChartboostSingleton.startChartboostRewardedVideo(context, mChartboostRewardedVideoDelegate);
   }
 
@@ -197,10 +204,11 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
         @Override
         public void didFailToLoadRewardedVideo(String location, CBError.CBImpressionError error) {
           super.didFailToLoadRewardedVideo(location, error);
+          String logMessage = ChartboostAdapterUtils.createSDKError(error);
+          Log.w(TAG, logMessage);
+
           if (mAdLoadCallback != null && location.equals(mChartboostParams.getLocation())) {
             if (mIsLoading) {
-              String logMessage = "Failed to load ad from Chartboost: " + error.toString();
-              Log.w(TAG, logMessage);
               mAdLoadCallback.onFailure(logMessage);
               mIsLoading = false;
             } else if (error == CBError.CBImpressionError.INTERNET_UNAVAILABLE_AT_SHOW) {
@@ -211,6 +219,16 @@ public class ChartboostMediationAdapter extends Adapter implements MediationRewa
                 mRewardedAdCallback.onAdFailedToShow("No network connection is available.");
               }
             }
+          }
+        }
+
+        @Override
+        public void onAdFailedToLoad(@AdapterError int errorCode, @NonNull String errorMessage) {
+          String logMessage = ChartboostAdapterUtils.createAdapterError(errorCode, errorMessage);
+          Log.w(TAG, logMessage);
+
+          if (mAdLoadCallback != null) {
+            mAdLoadCallback.onFailure(logMessage);
           }
         }